backend/pyproject.toml
[tool.poetry]
name = "config-blog"
version = "0.1.0"
description = ""
authors = ["plast7 <plast1206@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
